refactoring: moving the Sender out into the Azure package
